Added PA5: Garden
This is one of the largest programming assignments in the class. We need to create a simulation of a Garden using multiple classes and determine a meaningful organization for the classes on our own. Each plant in the Garden is of a specific type and has unique properties of its own. The attached pdf demonstrates the organization of the classes.